Abstract display methods and place in the UI class
