+++ src/main/java/seedu/exchangecoursemapper/parser/CourseValidator.java
@@ -2,19 +2,16 @@ package seedu.exchangecoursemapper.parser;
 
 import seedu.exchangecoursemapper.constants.Logs;
 import seedu.exchangecoursemapper.exception.Exception;
+import seedu.exchangecoursemapper.ui.UI;
 
 import javax.json.JsonArray;
 import javax.json.JsonObject;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import static seedu.exchangecoursemapper.constants.JsonKey.PU_COURSE_NAME_KEY;
 import static seedu.exchangecoursemapper.constants.JsonKey.PU_COURSE_CODE_KEY;
-import static seedu.exchangecoursemapper.constants.JsonKey.NUS_COURSE_NAME_KEY;
 import static seedu.exchangecoursemapper.constants.JsonKey.NUS_COURSE_CODE_KEY;
 import static seedu.exchangecoursemapper.constants.JsonKey.COURSES_ARRAY_LABEL;
-import static seedu.exchangecoursemapper.constants.Messages.LINE_SEPARATOR;
-import static seedu.exchangecoursemapper.constants.Messages.LIST_RELEVANT_PU;
 import static seedu.exchangecoursemapper.constants.Assertions.EMPTY_NUS_COURSE_WARNING;
 import static seedu.exchangecoursemapper.constants.Assertions.EMPTY_PU_WARNING;
 import static seedu.exchangecoursemapper.constants.Assertions.EMPTY_PU_COURSE_WARNING;
@@ -24,6 +21,7 @@ import static seedu.exchangecoursemapper.constants.Assertions.EMPTY_JSON_OBJECT_
 public class CourseValidator {
 
     private static final Logger logger = Logger.getLogger(CourseValidator.class.getName());
+    private UI ui = new UI();
 
     /**
      * Returns true if the user's course mapping is valid; otherwise, returns false.
@@ -59,32 +57,8 @@ public class CourseValidator {
         }
 
         System.out.println("Invalid course mapping!");
-        displayAvailableMappings(courses, pu);
+        ui.displayAvailableMappings(courses, pu);
         return false;
-    }
-
-    /**
-     * Prints all relevant course mappings for the user's PU input when an incorrect course mapping
-     * is passed from the isValidCourseMapping() method.
-     *
-     * @param courses a JsonArray containing all the relevant course mappings from the user's PU input.
-     * @param pu a string containing the user's PU input.
-     */
-    private void displayAvailableMappings(JsonArray courses, String pu) {
-        System.out.println("The available mappings for " + pu + " are:");
-        System.out.println(LINE_SEPARATOR);
-
-        for (int i = 0; i < courses.size(); i++) {
-            JsonObject course = courses.getJsonObject(i);
-            String puCourseCode = course.getString(PU_COURSE_CODE_KEY).toLowerCase();
-            String nusCourseCode = course.getString(NUS_COURSE_CODE_KEY).toLowerCase();
-            String nusCourseName = course.getString(NUS_COURSE_NAME_KEY).toLowerCase();
-            String puCourseName = course.getString(PU_COURSE_NAME_KEY).toLowerCase();
-
-            System.out.println(nusCourseCode + " " + nusCourseName + " | " + puCourseCode
-                    + " " + puCourseName + System.lineSeparator());
-        }
-        System.out.println(LINE_SEPARATOR);
     }
 
     /**
@@ -112,24 +86,12 @@ public class CourseValidator {
             logger.log(Level.INFO, Logs.RETRIEVE_COURSE_LIST);
             courses = jsonObject.getJsonObject(matchPu).getJsonArray(COURSES_ARRAY_LABEL);
         } else {
-            displayPartnerUniversities();
+            ui.displayPartnerUniversities();
             throw new IllegalArgumentException(Logs.INVALID_UNIVERSITY_INPUT);
         }
         return courses;
     }
 
-    /**
-     * Prints out all relevant partner universities available in our database.
-     */
-    private void displayPartnerUniversities() {
-        logger.log(Level.INFO, Logs.INVALID_UNIVERSITY_INPUT);
-        System.out.println(Logs.INVALID_UNIVERSITY_INPUT);
-
-        logger.log(Level.INFO, Logs.DISPLAY_PARTNER_UNIVERSITIES);
-        System.out.println(LINE_SEPARATOR);
-        System.out.println(LIST_RELEVANT_PU);
-        System.out.println(LINE_SEPARATOR);
-    }
 
     /**
      * Returns true if the course mapping is valid; otherwise, returns false.
